[#2570][#2559] refactor(web): Displaying a fallback while content is loading for each small module (#2574)
### What changes were proposed in this pull request?

1. Displaying a fallback while content is loading for the main content
Displaying a fallback while content is loading for metalakes list table
body module
Displaying a fallback while content is loading for catalog tree list
modul

![UILoading](https://github.com/datastrato/gravitino/assets/9210625/0911c818-ee1b-400d-9b58-3af764a59363)

2. Define metadata icon src path based on the operating environment

3. Fix some issue of inconsistent behaviour
- Smooth display of loaded data
- Preserve search parameters when strongly refreshing a page

![fixbugs](https://github.com/datastrato/gravitino/assets/9210625/dd94fce7-4ee0-4c76-a40e-94e4a87f5ecc)


### Why are the changes needed?

An instant loading state is fallback UI that is shown immediately upon
navigation.
You can pre-render loading indicators such as skeletons and spinners.
This helps users understand the app is responding and provides a better
user experience.

Fix issue because of the need for consistent behaviour after refactoring

Fix: #2570, #2559

### Does this PR introduce _any_ user-facing change?

N/A

### How was this patch tested?
Will run e2e test after PR #2534 merged

---------
