+++ web/src/app/metalakes/layout.js
@@ -7,28 +7,11 @@
 
 import { Grid, Typography } from '@mui/material'
 import { useSearchParams } from 'next/navigation'
-import { useEffect, useState } from 'react'
 import Metalake from './metalake/MetalakeView.js'
 
 const MetalakesView = ({ children }) => {
   const searchParams = useSearchParams()
-
-  const [, setRouteParams] = useState({})
-
   const metalakeName = searchParams.get('metalake')
-  const catalogName = searchParams.get('catalog')
-  const schemaName = searchParams.get('schema')
-  const tableName = searchParams.get('table')
-
-  useEffect(() => {
-    const takeParams = {
-      metalake: metalakeName,
-      catalog: catalogName,
-      schema: schemaName,
-      table: tableName
-    }
-    setRouteParams(takeParams)
-  }, [metalakeName, catalogName, schemaName, tableName])
 
   return (
     <>
