+++ web/src/app/metalakes/metalake/MetalakeView.js
@@ -29,33 +29,34 @@ const MetalakeView = () => {
   const dispatch = useAppDispatch()
   const searchParams = useSearchParams()
 
-  const routeParams = {
-    metalake: searchParams.get('metalake'),
-    catalog: searchParams.get('catalog'),
-    schema: searchParams.get('schema'),
-    table: searchParams.get('table')
-  }
+  const paramsSize = [...searchParams.keys()].length
 
   useEffect(() => {
+    const routeParams = {
+      metalake: searchParams.get('metalake'),
+      catalog: searchParams.get('catalog'),
+      schema: searchParams.get('schema'),
+      table: searchParams.get('table')
+    }
     if ([...searchParams.keys()].length) {
       const { metalake, catalog, schema, table } = routeParams
 
-      if (metalake) {
+      if (paramsSize === 1 && metalake) {
         dispatch(fetchCatalogs({ init: true, page: 'metalakes', metalake }))
         dispatch(getMetalakeDetails({ metalake }))
       }
 
-      if (catalog) {
+      if (paramsSize === 2 && catalog) {
         dispatch(fetchSchemas({ init: true, page: 'catalogs', metalake, catalog }))
         dispatch(getCatalogDetails({ metalake, catalog }))
       }
 
-      if (catalog && schema) {
+      if (paramsSize === 3 && catalog && schema) {
         dispatch(fetchTables({ init: true, page: 'schemas', metalake, catalog, schema }))
         dispatch(getSchemaDetails({ metalake, catalog, schema }))
       }
 
-      if (catalog && schema && table) {
+      if (paramsSize === 4 && catalog && schema && table) {
         dispatch(getTableDetails({ init: true, metalake, catalog, schema, table }))
       }
     }
@@ -73,7 +74,7 @@ const MetalakeView = () => {
     )
 
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [JSON.stringify(routeParams)])
+  }, [searchParams])
 
   return (
     <Box className={'metalake-template'} style={{ height: 'calc(100vh - 11rem)' }}>
